Swap UI of alerts and dashboards
