+++ src/server/routes/services/{serviceId}/diagnostics/dev/dashboards/promote/layout.njk
@@ -6,13 +6,13 @@
     {{ appPageHeading({
       caption: "Promote",
       text: "Alerts",
-      intro: 'Promote your playground alerts'
+      intro: 'Promote your playground dashboards'
     }) }}
 
     <div class="govuk-grid-row">
       <div class="govuk-grid-column-one-third-from-desktop-wide">
 
-      <p class="govuk-body govuk-!-margin-top-4">Select the alerts to be promoted to all environments.</p>
+      <p class="govuk-body govuk-!-margin-top-4">Select the dashboards to be promoted to all environments.</p>
 
       {% block formContent %}{% endblock %}
